Fixed auth manager and form d tests
